+++ pytest_play/engine.py
@@ -154,6 +154,7 @@ class PlayEngine(object):
                 'Command not implemented', command_type, provider_name)
         self.logger.info('Executing command %r', command)
         try:
+            print(command)
             return method(command, **kwargs)
         except Exception:
             self.logger.error('FAILED command %r', command)